New folder/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai-whisper
deep-translator
nrclex==3.0.0
edge-tts
python-dotenv==1.0.1
requests
